Put both emeter and usage module daymonthstat classes in single file (effecticely renamed usagestatus to daymonthstat) Put pydantic version of EmeterStatus in new module (emeterrealtime)
